backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
supabase==2.11.0
python-dotenv==1.0.1
websockets==14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pydantic[email]==2.10.3
python-multipart==0.0.19
